pyproject.toml
[project]
name = "dev-utils"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.1",
    "watchdog>=6.0.0",
]
